+++ NEWS.md
@@ -1,3 +1,21 @@
+Andromeda 1.1.0
+===============
+
+Changes:
+
+- Leveraging DuckDB for faster assigns and copy operations.
+
+- Added `flushAndromeda()` function.
+
+- Added `andromedaThreads` option to control the maximum number of threads Andromeda is allowed to use.
+
+Bugfixes:
+
+- Calling `flushAndromeda()` before copying entire Andromeda (`copyAndromeda()`) or just a table (`[[]]<-` operator) to avoid segfault.
+
+- Switch from `zip::unzip()` to `utils::unzip()` to avoid 'mtimes' errors.
+
+
 Andromeda 1.0.0
 ===============